smart_parse/requirements.txt
openai
grpcio
grpcio-tools